Separate DeviceHelper Class
