+++ Assets/Scripts/Devices/DeviceHelper.Camera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 432c6df1da39bd87ba0781f8cbc5ffd0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
